+++ app/src/main/java/pk/sufiishq/app/feature/personalize/controller/PersonalizeController.kt
@@ -14,13 +14,16 @@
  * limitations under the License.
  */
 
-package pk.sufiishq.app.utils
+package pk.sufiishq.app.feature.personalize.controller
 
-const val SUFI_ISHQ_HOST = "https://www.sufiishq.pk/"
-const val KALAM_DIR = "kalam"
-const val MEDIA_PATH = "media"
-const val CACHE_SPLIT_FILENAME = "split.mp3"
-const val KALAM_TITLE_LENGTH = 50
-const val PLAYLIST_TITLE_LENGTH = 30
-const val LAST_PLAY_KALAM = "last_play_kalam"
-const val FEATURE_VIDEO_REFERENCE_ID = "0"
+import androidx.lifecycle.LiveData
+import pk.sufiishq.app.feature.personalize.model.LogoPath
+import pk.sufiishq.app.feature.personalize.model.Personalize
+
+interface PersonalizeController {
+
+    suspend fun resolveLogo(logoPath: LogoPath): LogoPath?
+    fun get(): LiveData<Personalize?>
+    fun reset()
+    fun update(personalize: Personalize)
+}