+++ app/src/main/java/pk/sufiishq/app/feature/personalize/model/Personalize.kt
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package pk.sufiishq.app.utils
+package pk.sufiishq.app.feature.personalize.model
 
-const val SUFI_ISHQ_HOST = "https://www.sufiishq.pk/"
-const val KALAM_DIR = "kalam"
-const val MEDIA_PATH = "media"
-const val CACHE_SPLIT_FILENAME = "split.mp3"
-const val KALAM_TITLE_LENGTH = 50
-const val PLAYLIST_TITLE_LENGTH = 30
-const val LAST_PLAY_KALAM = "last_play_kalam"
-const val FEATURE_VIDEO_REFERENCE_ID = "0"
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "personalize")
+data class Personalize(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int,
+    val frontPath: String,
+    val backPath: String,
+)